requirements.txt
altair==5.2.0
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
debtcollector==2.5.0
distlib==0.3.8
Django==4.2.8
dnspython==2.4.2
docopt==0.6.2
eventlet==0.34.2
fasteners==0.19
filelock==3.13.1
Flask==3.0.0
gitdb==4.0.11
GitPython==3.1.40
googletrans==4.0.0rc1
greenlet==3.0.3
gTTS==2.5.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
image==1.5.33
importlib-metadata==6.11.0
importlib-resources==6.1.1
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
Js2Py==0.74
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgpack==1.0.7
netaddr==0.9.0
netifaces==0.11.0
nltk==3.8.1
numpy==1.24.4
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
os-win==5.9.0
oslo.concurrency==5.2.0
oslo.config==9.2.0
oslo.context==5.3.0
oslo.i18n==6.2.0
oslo.log==5.4.0
oslo.serialization==5.2.0
oslo.utils==6.3.0
packaging==23.2
pandas==2.0.3
pbr==6.0.0
Pillow==10.1.0
pipwin==0.5.2
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
playsound==1.3.0
protobuf==4.25.1
pyarrow==14.0.2
PyAudio @ file:///C:/Users/aadya/pipwin/PyAudio-0.2.11-cp38-cp38-win_amd64.whl#sha256=4cee0b30bbb8bf5b7dee95c6f010e55bf80024348b35295dc5cc86be3c274e22
pydeck==0.8.1b0
Pygments==2.17.2
pyjsparser==2.7.1
PyMI==1.0.6
pyparsing==3.1.1
PyPrind==2.11.3
pySmartDL==1.3.4
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.15.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SpeechRecognition==3.10.1
sqlparse==0.4.4
stevedore==5.1.0
streamlit==1.29.0
tenacity==8.2.3
tesseract==0.1.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
virtualenv==20.25.0
watchdog==3.0.0
Werkzeug==3.0.1
WMI==1.5.1
wrapt==1.16.0
zipp==3.17.0
